chapter-42-research-assistant/code/requirements.txt
# Requirements for Chapter 42: Research Assistant Agent

# Anthropic SDK for Claude API
anthropic>=0.40.0

# Web requests
requests>=2.31.0

# HTML parsing for web_read_tool
beautifulsoup4>=4.12.0

# Environment variable management
python-dotenv>=1.0.0
